frontend/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
anyio==3.6.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cycler==0.11.0
datasets==2.6.1
debugpy==1.5.1
decorator==5.1.1
dill==0.3.5.1
donut-python==1.0.8
entrypoints==0.4
fastapi==0.86.0
filelock==3.8.0
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.15.2
ipython==7.31.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter_client==7.3.5
jupyter_core==4.11.1
kiwisolver==1.4.4
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.2
multiprocess==0.70.13
munch==2.5.0
nest-asyncio==1.5.5
nltk==3.7
numpy==1.21.6
oauthlib==3.2.2
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.20
protobuf==3.19.6
psutil==5.9.0
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytorch-lightning==1.8.0.post1
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==302
PyYAML==6.0
pyzmq==23.2.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sconf==0.2.5
semver==2.13.0
sentencepiece==0.1.97
setuptools==65.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.20.4
streamlit==1.14.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
timm==0.6.11
tokenizers==0.13.1
toml==0.10.2
toolz==0.12.0
torch==1.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
transformers==4.25.0.dev0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.19.0
validators==0.20.0
watchdog==2.1.9
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
xxhash==3.1.0
yapf==0.32.0
yarl==1.8.1
zipp==3.10.0
zss==1.2.0
